+++ src/data/nutritional_information/FoodTextData.ts
